Feat : 카드 업데이트,삭제시 조건문 추가
servletRequest를 받아서 유저 id를 조회하여 작성자의 id와 현제 접속중이 유저의 id가 다르면 오류가 출력되도록 변경

+++ src/main/java/com/example/trello/card/CardService.java
@@ -49,10 +49,16 @@ public class CardService {
     }
 
     //카드 업데이트
-    public CardResponseDto updateCardService(Long cardId, UpdateCardRequestDto requestDto) {
+    public CardResponseDto updateCardService(Long cardId, UpdateCardRequestDto requestDto, HttpServletRequest servletRequest) {
         Card card = cardRepository.findByIdOrElseThrow(cardId);
 
         CardList cardList = cardListRepository.findByIdOrElseThrow(requestDto.getCardListId());
+
+        User user = (User) servletRequest.getSession().getAttribute("id");
+
+        if (!user.getId().equals(cardId)) {
+            throw new RuntimeException();
+        }
 
         card.updateCard(cardList, requestDto.getTitle(), requestDto.getDescription(), requestDto.getStartAt(), requestDto.getEndAt());
 
@@ -63,8 +69,14 @@ public class CardService {
 
     // 카드 삭제
     @PostMapping("/{cardsId}")
-    public void deleteCardService(Long cardId) {
+    public void deleteCardService(Long cardId, HttpServletRequest servletRequest) {
         Card card = cardRepository.findByIdOrElseThrow(cardId);
+
+        User user = (User) servletRequest.getSession().getAttribute("id");
+
+        if (!user.getId().equals(cardId)) {
+            throw new RuntimeException();
+        }
         cardRepository.delete(card);
     }
